Add files via upload
Please use these scripts for FREE LICENSE and distribution among DBA's to Alert about a very special case where Application leaves the Session in "Sleeping" state with ACTIVE LOCKS.

SQL server does not have problem with Sleeping sessions, there are many Application that keeps the Session open and run different batches/Query within the same session for Days and that's OK with SQL server. It has been designed to handle such sessions, however, if the Application keeps the Session's sleeping with ACTIVE LOCKS than that becomes a problem. There are many LOCK types which are not compatible with each other, resulting SQL server starts reporting BLOCKING and it hangs the Application.

To report and KILL such sessions, please use the Script "as-is" and it will start reporting those sessions to defined EMAILS and can even KILL TOO if instructed in Scripts.